+++ src/Realm.Domain/Registries/ItemRegistryEntry.cs
@@ -9,7 +9,7 @@ public class ItemRegistryEntry
     }
 
     public uint Id { get; set; }
-    public uint Size { get; set; }
+    public decimal Size { get; set; }
     public string Name { get; set; }
     public uint StackSize { get; set; } = 1; // 1 = not stackable
     public ItemAction AvailiableActions { get; set; }
